+++ tests/Metalama.Framework.Tests.Integration/Tests/Aspects/Introductions/Events/DesignTime.0.i.cs
@@ -1,0 +1,16 @@
+namespace Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Events.DesignTime
+{
+  partial class TargetClass
+  {
+    public event global::System.EventHandler? Event
+    {
+      add
+      {
+      }
+      remove
+      {
+      }
+    }
+    public event global::System.EventHandler? EventField;
+  }
+}
