+++ tests/Metalama.Framework.Tests.Integration/Tests/Aspects/Introductions/Interfaces/DesignTimeExplicitMembers.0.i.cs
@@ -1,0 +1,39 @@
+namespace Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers
+{
+  partial class TargetClass : global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface
+  {
+    global::System.Int32 global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface.InterfaceMethod()
+    {
+      return default(global::System.Int32);
+    }
+    global::System.Int32 global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface.AutoProperty { get; set; }
+    global::System.Int32 global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface.Property
+    {
+      get
+      {
+        return default(global::System.Int32);
+      }
+      set
+      {
+      }
+    }
+    event global::System.EventHandler global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface.EventField
+    {
+      add
+      {
+      }
+      remove
+      {
+      }
+    }
+    event global::System.EventHandler global::Metalama.Framework.Tests.Integration.TestInputs.Aspects.Introductions.Interfaces.DesignTimeExplicitMembers.IInterface.Event
+    {
+      add
+      {
+      }
+      remove
+      {
+      }
+    }
+  }
+}
