+++ tests/Metalama.Framework.Tests.Integration/Tests/Aspects/Introductions/Methods/DesignTime.0.i.cs
@@ -1,0 +1,25 @@
+namespace Metalama.Framework.IntegrationTests.Aspects.Introductions.Methods.DesignTime
+{
+  partial class TargetClass
+  {
+    public global::System.Int32 IntroducedMethod_Int()
+    {
+      return default(global::System.Int32);
+    }
+    public global::System.Int32 IntroducedMethod_Param(global::System.Int32 x)
+    {
+      return default(global::System.Int32);
+    }
+    public static global::System.Int32 IntroducedMethod_StaticSignature()
+    {
+      return default(global::System.Int32);
+    }
+    public virtual global::System.Int32 IntroducedMethod_VirtualExplicit()
+    {
+      return default(global::System.Int32);
+    }
+    public void IntroducedMethod_Void()
+    {
+    }
+  }
+}
