+++ tests/Metalama.Framework.Tests.Integration/Tests/Aspects/Introductions/Properties/DesignTime.0.i.cs
@@ -1,0 +1,16 @@
+namespace Metalama.Framework.IntegrationTests.Aspects.Introductions.Properties.DesignTime
+{
+  partial class TargetClass
+  {
+    public global::System.Int32 IntroducedProperty_Accessors
+    {
+      get
+      {
+        return default(global::System.Int32);
+      }
+      set
+      {
+      }
+    }
+  }
+}
