+++ tests/Metalama.Framework.Tests.Integration/Tests/Aspects/Introductions/Properties/DesignTime.t.cs
@@ -1,16 +1,4 @@
-namespace Metalama.Framework.IntegrationTests.Aspects.Introductions.Properties.DesignTime
+[Introduction]
+internal partial class TargetClass
 {
-  partial class TargetClass
-  {
-    public global::System.Int32 IntroducedProperty_Accessors
-    {
-      get
-      {
-        return default(global::System.Int32);
-      }
-      set
-      {
-      }
-    }
-  }
 }
